backend/src/test_case_gen/requirements.txt
functions-framework==3.9.2
google-genai==1.38.0
libsql==0.1.11